+++ core/src/main/kotlin/com/gitlab/kordlib/core/behavior/WebhookBehavior.kt
@@ -1,12 +1,12 @@
 package com.gitlab.kordlib.core.behavior
 
+import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.Kord
-import com.gitlab.kordlib.rest.builder.webhook.ExecuteWebhookBuilder
-import com.gitlab.kordlib.rest.builder.webhook.WebhookModifyBuilder
 import com.gitlab.kordlib.core.cache.data.WebhookData
 import com.gitlab.kordlib.core.entity.Entity
-import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.entity.Webhook
+import com.gitlab.kordlib.rest.builder.webhook.ExecuteWebhookBuilder
+import com.gitlab.kordlib.rest.builder.webhook.WebhookModifyBuilder
 
 /**
  * The behavior of a [Discord Webhook](https://discordapp.com/developers/docs/resources/webhook).
@@ -28,7 +28,7 @@ interface WebhookBehavior : Entity {
     }
 
     companion object {
-        internal operator fun invoke(id: Snowflake, kord: Kord) : WebhookBehavior = object : WebhookBehavior {
+        internal operator fun invoke(id: Snowflake, kord: Kord): WebhookBehavior = object : WebhookBehavior {
             override val id: Snowflake = id
             override val kord: Kord = kord
         }
@@ -43,11 +43,7 @@ interface WebhookBehavior : Entity {
  */
 @Suppress("NAME_SHADOWING")
 suspend inline fun WebhookBehavior.edit(builder: WebhookModifyBuilder.() -> Unit): Webhook {
-    val builder = WebhookModifyBuilder().apply(builder)
-    val reason = builder.reason
-    val request = builder.toRequest()
-
-    val response = kord.rest.webhook.modifyWebhook(id.value, request, reason)
+    val response = kord.rest.webhook.modifyWebhook(id.value, builder)
     val data = WebhookData.from(response)
 
     return Webhook(data, kord)
@@ -60,11 +56,7 @@ suspend inline fun WebhookBehavior.edit(builder: WebhookModifyBuilder.() -> Unit
  */
 @Suppress("NAME_SHADOWING")
 suspend inline fun WebhookBehavior.edit(token: String, builder: WebhookModifyBuilder.() -> Unit): Webhook {
-    val builder = WebhookModifyBuilder().apply(builder)
-    val reason = builder.reason
-    val request = builder.toRequest()
-
-    val response = kord.rest.webhook.modifyWebhookWithToken(id.value, token, request, reason)
+    val response = kord.rest.webhook.modifyWebhookWithToken(id.value, token, builder)
     val data = WebhookData.from(response)
 
     return Webhook(data, kord)
@@ -74,6 +66,5 @@ suspend inline fun WebhookBehavior.edit(token: String, builder: WebhookModifyBui
  * Requests to execute this webhook.
  */
 suspend inline fun WebhookBehavior.execute(token: String, builder: ExecuteWebhookBuilder.() -> Unit) {
-    val request = ExecuteWebhookBuilder().apply(builder).toRequest()
-    kord.rest.webhook.executeWebhook(token = token, webhookId = id.value, request = request, wait = true)
+    kord.rest.webhook.executeWebhook(token = token, webhookId = id.value, wait = true, builder = builder)
 }