+++ core/src/main/kotlin/com/gitlab/kordlib/core/behavior/channel/CategoryBehavior.kt
@@ -7,6 +7,7 @@ import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.entity.channel.CategorizableChannel
 import com.gitlab.kordlib.core.entity.channel.Category
 import com.gitlab.kordlib.core.entity.channel.Channel
+import com.gitlab.kordlib.rest.service.patchCategory
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.filterIsInstance
@@ -37,11 +38,7 @@ interface CategoryBehavior : GuildChannelBehavior {
  */
 @Suppress("NAME_SHADOWING")
 suspend fun CategoryBehavior.edit(builder: CategoryModifyBuilder.() -> Unit): Category {
-    val builder = CategoryModifyBuilder().apply(builder)
-    val reason = builder.reason
-    val request = builder.toRequest()
-
-    val response = kord.rest.channel.patchChannel(id.value, request, reason)
+    val response = kord.rest.channel.patchCategory(id.value, builder)
     val data = ChannelData.from(response)
 
     return Channel.from(data, kord) as Category