+++ core/src/main/kotlin/com/gitlab/kordlib/core/behavior/channel/StoreChannelBehavior.kt
@@ -1,11 +1,12 @@
 package com.gitlab.kordlib.core.behavior.channel
 
 import com.gitlab.kordlib.core.Kord
-import com.gitlab.kordlib.rest.builder.channel.UpdateStoreChannelBuilder
+import com.gitlab.kordlib.rest.builder.channel.StoreChannelModifyBuilder
 import com.gitlab.kordlib.core.cache.data.ChannelData
 import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.entity.channel.Channel
 import com.gitlab.kordlib.core.entity.channel.StoreChannel
+import com.gitlab.kordlib.rest.service.patchStoreChannel
 
 /**
  * The behavior of a Discord Store Channel associated to a guild.
@@ -31,13 +32,8 @@ interface StoreChannelBehavior : GuildChannelBehavior {
  *
  * @return The edited [StoreChannel].
  */
-@Suppress("NAME_SHADOWING")
-suspend inline fun StoreChannelBehavior.edit(builder: (UpdateStoreChannelBuilder) -> Unit): StoreChannel {
-    val builder = UpdateStoreChannelBuilder().apply(builder)
-    val reason = builder.reason
-    val request = builder.toRequest()
-
-    val response = kord.rest.channel.patchChannel(id.value, request, reason)
+suspend inline fun StoreChannelBehavior.edit(builder: StoreChannelModifyBuilder.() -> Unit): StoreChannel {
+    val response = kord.rest.channel.patchStoreChannel(id.value, builder)
     val data = ChannelData.from(response)
 
     return Channel.from(data, kord) as StoreChannel