+++ core/src/main/kotlin/com/gitlab/kordlib/core/behavior/channel/TextChannelBehavior.kt
@@ -6,6 +6,7 @@ import com.gitlab.kordlib.core.cache.data.ChannelData
 import com.gitlab.kordlib.common.entity.Snowflake
 import com.gitlab.kordlib.core.entity.channel.Channel
 import com.gitlab.kordlib.core.entity.channel.TextChannel
+import com.gitlab.kordlib.rest.service.patchTextChannel
 
 
 interface TextChannelBehavior : GuildMessageChannelBehavior {
@@ -29,14 +30,8 @@ interface TextChannelBehavior : GuildMessageChannelBehavior {
  *
  * @return The edited [TextChannel].
  */
-
-@Suppress("NAME_SHADOWING")
-suspend inline fun TextChannelBehavior.edit(builder: (TextChannelModifyBuilder) -> Unit): TextChannel {
-    val builder = TextChannelModifyBuilder().apply(builder)
-    val reason = builder.reason
-
-    val request = builder.toRequest()
-    val response = kord.rest.channel.patchChannel(id.value, request, reason)
+suspend inline fun TextChannelBehavior.edit(builder: TextChannelModifyBuilder.() -> Unit): TextChannel {
+    val response = kord.rest.channel.patchTextChannel(id.value, builder)
 
     val data = ChannelData.from(response)
     return Channel.from(data, kord) as TextChannel